feat(WIP): Sessions can now be run in a Windows Service context
Problem:

Windows requires the use of different Win32 APIs to impersonate users when
you're running in the context of a Service (Windows Session ID 0) than if
you're running in an interactive logon session. The current code uses
CreateProcessWithLogonW to do impersonation, and that API does not work inside Session 0.

Solution:

Detect when we are running in Session 0, and make use of CreateProcessAsUserW instead
in that case. To support this we require that when running in Session 0, the caller of
this library create the WindowsSessionUser using the handle to a logon token that they
create with the Win32 APIs rather than with a password directly.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>